feat(db): add downloading scripts for various databases

+++ src/1_gathering/db/afrotryp/downloading_modified.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
 # -*- coding: utf-8 -*-
 
-mkdir -p ../data/external/dbSource/afrotryp/
+if [ ! -f ../data/external/dbSource/afrotryp/afrotryp.tsv.zip ]; then
 
-wget "https://osf.io/q47fd/download" -O ../data/external/dbSource/afrotryp/afrotryp.tsv.zip
+	mkdir -p ../data/external/dbSource/afrotryp/
+
+	wget "https://osf.io/q47fd/download" -O ../data/external/dbSource/afrotryp/afrotryp.tsv.zip
+fi
+
+echo "Done"
