+++ src/1_gathering/db/tmmc/downloading.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash
 # -*- coding: utf-8 -*-
 
+if [ ! -f ../data/external/dbSource/tmmc/compound.xlsx ]; then
+
 mkdir -p ../data/external/dbSource/tmmc/
 
 wget "http://informatics.kiom.re.kr/compound/download/compound.xlsx" -O ../data/external/dbSource/tmmc/compound.xlsx
+
+fi
+
+echo "Done"
